requirements.txt
streamlit
chromadb
google-genai
python-dotenv
slack-sdk
google-auth
google-auth-oauthlib
google-api-python-client